requirements.txt
Markdown==2.6.11
pkg-resources==0.0.0
python-frontmatter==0.4.2
PyYAML==3.13
six==1.11.0
